add timer task

+++ include/system/register.h
@@ -48,9 +48,9 @@ typedef struct {
 
 void register_init();
 
-void register_write_byte(uint16_t addr, uint8_t value);
-void register_set_bit(uint16_t addr, BIT_X bit);
-void register_clear_bit(uint16_t addr, BIT_X bit);
+void register_write_byte(int addr, uint8_t value);
+void register_set_bit(int addr, BIT_X bit);
+void register_clear_bit(int addr, BIT_X bit);
 
 void register_port_write_byte(PORT_X port, uint8_t value);
 void register_ddr_write_byte(PORT_X port, uint8_t value);
